Set up Vuex store, mostly refactored/componetized, minor ui tweaks
