+++ package.json
@@ -10,7 +10,8 @@
   "dependencies": {
     "axios": "^0.18.0",
     "simplebar-vue": "^0.1.1",
-    "vue": "^2.5.17"
+    "vue": "^2.5.17",
+    "vuex": "^3.0.1"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "^3.0.5",
